add yml2db.py

+++ sources/extract.sh
@@ -6,7 +6,12 @@ archives=$(find ../archives/ -name "${project}*.tar.gz")
 
 for archive in $archives; do
   basename=$(basename $archive .tar.gz)
-  tar xvf $archive
-  mv ${project} $basename
+  if [ ! -e $basename ]; then
+    echo "extract $archive ..."
+    tar xf $archive
+    mv ${project} $basename
+  else
+    echo "skip extracting $archive"
+  fi
 done
